requirements.txt
asgiref==3.5.2
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
pandas==2.0.1
regex==2023.5.5
jaro-winkler==2.0.3
sentence-transformers==2.2.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.0